Major refactor to use vulkan context singleton
This change abstracts a lot of the "static" Vulkan lifetime data
structures (such as instance, device, swapchain, etc.) to be accessed
through a singleton `VulkanContext`. This is probably not the best
solution for multi-threaded code, but simplifies a large portion of code
by not having to pass device and allocator references to all parts of
the renderer.